update(file): file structure is new, Makefile is improved and worflow is more efficient
